move nameserver codes into agent
