+++ agent/nameserver/resolver.go
