+++ agent/nameserver/type.go
